% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{AUC}
\alias{AUC}
\title{Calculate AUC}
\usage{
AUC(xb.hat, y)
}
\arguments{
\item{xb.hat}{A vector with the predictions}

\item{y}{A vector with the observed outcomes}
}
\description{
Calculates AUC for binary outcomes
}
\author{
David van Klaveren
}
