% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{AUC.trinary}
\alias{AUC.trinary}
\title{Calculate AUC}
\usage{
AUC.trinary(xb.hat, y)
}
\arguments{
\item{xb.hat}{A vector with the predictions}

\item{y}{A vector with the observed outcomes}
}
\description{
Calculates AUC for trinary outcomes like the ones generated from matching
patients
}
\author{
David van Klaveren
}
