% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{calculateCalibrationForBenefit}
\alias{calculateCalibrationForBenefit}
\title{Calculate the calibration for benefit}
\usage{
calculateCalibrationForBenefit(data, method = "rank")
}
\arguments{
\item{data}{A dataframe with columns \code{treatment}, \code{outcome} and
\code{predictedBenefit}}

\item{method}{The method for matching. Currently, only \code{rank} is supported}
}
\description{
Calculates the ICI, E50 and E90 for benefit, using matched
patient pairs
}
\author{
Carolien Maas
}
