% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{calculateCForBenefit}
\alias{calculateCForBenefit}
\title{Calculate the C-statistic for benefit}
\usage{
calculateCForBenefit(data, method = "rank")
}
\arguments{
\item{data}{A dataframe with columns ...}

\item{method}{Can be "rank" ....}

\item{smoothFit}{A smooth model for predicting absolute benefit}
}
\description{
Calculates the C-statistic for benefit for the evaluation of
discrimination
}
