% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createAdaptiveSettings}
\alias{createAdaptiveSettings}
\title{Create adaptive settings}
\usage{
createAdaptiveSettings(settings)
}
\arguments{
\item{settings}{A list of settings}
}
\description{
Creates adaptive settings
}
