% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createHteSettings}
\alias{createHteSettings}
\title{Create settings}
\usage{
createHteSettings(label, settings)
}
\arguments{
\item{label}{The label of the fitted method. Useful if multiple methods
are considered}

\item{settings}{The settings for fitting the method}
}
\description{
Create the settings for a specific approach to treatment effect heterogeneity
}
