% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createLoessSettings}
\alias{createLoessSettings}
\title{Create Loess Settings}
\usage{
createLoessSettings(
  weights = numeric(),
  model = FALSE,
  span = 0.75,
  degree = 2,
  parametric = FALSE,
  drop.square = FALSE,
  normalize = TRUE,
  family = "gaussian",
  method = "loess",
  control = stats::loess.control()
)
}
\arguments{
\item{weights}{Optional weights for each case.}

\item{model}{Should the model frame be returned?}

\item{span}{The parameter alpha which controls the degree of smoothing.}

\item{degree}{the degree of the polynomials to be used, normally 1 or 2.
(Degree 0 is also allowed, but see the ‘Note’.)}

\item{parametric}{Should any terms be fitted globally rather than locally?
vector of the same length as the number of predictors.}

\item{drop.square}{For fits with more than one predictor and degree = 2,
should the quadratic term be dropped for particular
predictors? Terms are specified in the same way as for
parametric.}

\item{normalize}{Should the predictors be normalized to a common scale if
there is more than one? The normalization used is to set
the 10\% trimmed standard deviation to one. Set to false
for spatial coordinate predictors and others known to be
on a common scale.}

\item{family}{If "gaussian" fitting is by least-squares, and if
"symmetric" a re-descending M estimator is used with
Tukey's biweight function. Can be abbreviated.}

\item{method}{Fit the model or just extract the model frame. Can be
abbreviated.}

\item{control}{Control parameters.}
}
\description{
Creates the settings for fitting a loess fit.
}
