% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{createPairs}
\alias{createPairs}
\title{Create matched pairs}
\usage{
createPairs(data)
}
\arguments{
\item{data}{A dataframe with columns \code{treatment}, \code{outcome} and
\code{predictedBenefit}}
}
\description{
Creates matched pairs for the calculation of discrimination and
calibration for benefit
}
\author{
Carolien Maas
}
