% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createStratifiedSettings}
\alias{createStratifiedSettings}
\title{Create stratified analysis settings}
\usage{
createStratifiedSettings(nStrata = 4)
}
\arguments{
\item{nStrata}{The number of strata in which the population will be
divided. Currently, only equal-sized risk strata are
supported.}
}
\description{
Creates the settings for fitting a risk-stratified analysis
}
