% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSmooth.R
\name{fitAdaptive}
\alias{fitAdaptive}
\title{Fit adaptive approach}
\usage{
fitAdaptive(data, settings)
}
\arguments{
\item{data}{The data on which the adaptive approach will be applied.
Requires a column called \code{riskLinearPredictor} with the
linear predictors of the existing prediction model}

\item{settings}{A list of settings for making smooth benefit predictions}
}
\description{
Fits an adaptive approach for assessing treatment effect heterogeneity
}
