% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSmooth.R
\name{fitLocfitHte}
\alias{fitLocfitHte}
\title{Fit Restricted Cubic Spline HTE}
\usage{
fitLocfitHte(data, settings)
}
\arguments{
\item{data}{A dataframe containing at least a column named
"riskLinearPredictor" and column named "outcome".
Currently, only binary outcomes are supported.}

\item{settings}{A list of settings generated from \code{\link[=createLocfitSettings]{createLocfitSettings()}}}
}
\description{
Fit a local likelihood model for risk-based smoothing of absolute benefit
}
