% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSmooth.R
\name{fitLoessHte}
\alias{fitLoessHte}
\title{Fit Loess HTE}
\usage{
fitLoessHte(data, loessSettings)
}
\arguments{
\item{data}{A dataframe containing at least a column named
"riskLinearPredictor" and column named "outcome".
Currently, only binary outcomes are supported.}

\item{loessSettings}{A list of settings generated from \code{\link[=createLoessSettings]{createLoessSettings()}}}
}
\description{
Fit a loess type of model to a dataframe
}
