% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSmooth.R
\name{fitModelBasedHte}
\alias{fitModelBasedHte}
\title{Fit model-based heterogeneity of treatment effect}
\usage{
fitModelBasedHte(data, settings)
}
\arguments{
\item{data}{A dataframe containing at least a column named
"riskLinearPredictor" and column named "outcome".
Currently, only binary outcomes are supported.}

\item{settings}{A list of settings generated from \code{\link[=createModelBasedSettings]{createModelBasedSettings()}}}
}
\description{
Fits a model-based approach to treatment effect heterogeneity
}
