% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSmooth.R
\name{fitRcsHte}
\alias{fitRcsHte}
\title{Fit Restricted Cubic Spline HTE}
\usage{
fitRcsHte(data, settings)
}
\arguments{
\item{data}{A dataframe containing at least a column named
"riskLinearPredictor" and column named "outcome".
Currently, only binary outcomes are supported.}

\item{settings}{A list of settings generated from \code{\link[=createRcsSettings]{createRcsSettings()}}}
}
\description{
Fit a restricted cubic spline smooth type of model to a dataframe
}
