% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSmooth.R
\name{fitStratifiedHte}
\alias{fitStratifiedHte}
\title{Fit risk-stratified heterogeneity of treatment effect}
\usage{
fitStratifiedHte(data, settings)
}
\arguments{
\item{data}{A dataframe containing at least a column named
"riskLinearPredictor" and column named "outcome".
Currently, only binary outcomes are supported.}

\item{settings}{A list of settings generated from \code{\link[=createStratifiedSettings]{createStratifiedSettings()}}}
}
\description{
Fits a risk-stratified approach for assessing treatment effect heterogeneity
}
