% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSmooth.R
\name{predictBenefitLoess}
\alias{predictBenefitLoess}
\title{Predict LOESS benefit}
\usage{
predictBenefitLoess(p, s0, s1)
}
\arguments{
\item{p}{The baseline risk probability}

\item{s0}{The fit in the untreated arm}

\item{s1}{The fit in the treated arm}
}
\description{
Predicts absolute benefit based on two loess fits, one in each treatment arm
}
