% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSmooth.R
\name{predictBenefitModelBasedHte}
\alias{predictBenefitModelBasedHte}
\title{Predict model-based treatment effects}
\usage{
predictBenefitModelBasedHte(p, modelBasedFit)
}
\arguments{
\item{p}{The new data on which the prediction is made}

\item{modelBasedFit}{The estimated model-based fit. See \code{\link[=fitModelBasedHte]{fitModelBasedHte()}}}
}
\description{
Predicts absolute benefit using the model-based approach to treatment effect
heterogeneity
}
