% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSmooth.R
\name{predictSmoothBenefit}
\alias{predictSmoothBenefit}
\title{Predict smooth benefit}
\usage{
predictSmoothBenefit(p, smoothFit)
}
\arguments{
\item{p}{The new data on which the prediction is made}

\item{smoothControl}{The smooth fit in the control arm}

\item{smoothTreatment}{The smooth fit in the treatment arm}
}
\description{
Predicts the absolute benefit based on two smooth fits
}
