% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSmooth.R
\name{predictStratifiedBenefit}
\alias{predictStratifiedBenefit}
\title{Predict risk-stratified benefits}
\usage{
predictStratifiedBenefit(p, stratifiedHte)
}
\arguments{
\item{p}{The new data on which the prediction is made}

\item{stratifiedHte}{The estimated risk-stratified fit. See \code{\link[=fitStratifiedHte]{fitStratifiedHte()}}}
}
\description{
Predicts absolute benefit using the risk-stratified approach to treatment
effect heterogeneity
}
