% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{calculateCForBenefit}
\alias{calculateCForBenefit}
\title{Calculate the C-statistic for benefit}
\usage{
calculateCForBenefit(data)
}
\arguments{
\item{data}{A dataframe with columns ...}
}
\description{
Calculates the C-statistic for benefit for the evaluation of
discrimination
}
