% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createLoessSettings}
\alias{createLoessSettings}
\title{Create Loess Settings}
\usage{
createLoessSettings(span = 0.5, degree = 1)
}
\arguments{
\item{span}{The parameter alpha which controls the degree of smoothing.}

\item{degree}{the degree of the polynomials to be used, normally 1 or 2.
(Degree 0 is also allowed, but see the ‘Note’.)}
}
\description{
Creates the settings for fitting a loess fit.
}
