% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{createPairs}
\alias{createPairs}
\title{Create matched pairs}
\usage{
createPairs(data, method = "rank")
}
\arguments{
\item{data}{A dataframe with columns \code{treatment}, \code{outcome} and
\code{predictedBenefit}}

\item{method}{The method for matching. Currently, only \code{rank} is supported}
}
\description{
Creates matched pairs for the calculation of discrimination and
calibration for benefit
}
