% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSmooth.R
\name{predictSmooth}
\alias{predictSmooth}
\title{Predict with smooth fit}
\usage{
predictSmooth(smoothFit, p)
}
\arguments{
\item{smoothFit}{An estimated smooth fit}

\item{p}{The new data on which the prediction is made}
}
\description{
Predict individualized benefit for new data based on a smooth fit.
}
